requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
pydantic
bcrypt
python-jose[cryptography]
python-dotenv
slowapi
